♻️ [Refactor] common 디렉토리 생성
